correct a minor issue

+++ src/Mesh/ElemBC_3D_tet_outflow.cpp
@@ -6,7 +6,7 @@ ElemBC_3D_tet_outflow::ElemBC_3D_tet_outflow(
     const int &elemtype )
 : ElemBC_3D_tet( vtkfileList, elemtype )
 {
-  SYS_T::print_fatal_if(outlet_normal_vec.size() != num_ebc,
+  SYS_T::print_fatal_if(outlet_normal_vec.size() != static_cast<unsigned int>( num_ebc ),
       "Error: the input normal vector length does not match the number of outlets.\n");
 
   outNormal.resize( num_ebc );
